+++ packages/client/src/index.ts
@@ -27,7 +27,6 @@ export * from './base/drag-aware-mouse-listener';
 export * from './base/editor-context-service';
 export * from './base/feedback/css-feedback';
 export * from './base/feedback/feedback-action-dispatcher';
-export * from './base/feedback/feedback-action-dispatcher-default';
 export * from './base/feedback/feedback-command';
 export * from './base/feedback/feedback-emitter';
 export * from './base/feedback/set-model-command';
@@ -145,7 +144,6 @@ export * from './features/grid/grid-snapper';
 export * from './features/grid/grid-style';
 export * from './features/helper-lines/helper-line-feedback';
 export * from './features/helper-lines/helper-line-manager';
-export * from './features/helper-lines/helper-line-manager-default';
 export * from './features/helper-lines/helper-line-module';
 export * from './features/helper-lines/model';
 export * from './features/helper-lines/view';
